Initial commit: basic framework and instances of deal_damage
Every instance of deal_damage, has, where applicable, been given a source and arguments for whether it is forced and trackable.
In many cases I have chosen to force the damage: for example, in status effects or very frequent periodic damage, or things you really shouldn't be reacting to with a parry or similar (like WCCA trapping you inside)
In some cases I have chosen to make the damage untrackable, which means it will not add aggro to hostile mobs hit by it, for example TSO, since we don't want mobs to track it down for players.